Made data members of Variable protected rather than private. Renamed Expression::evaluatesToFloat() and Expression::evaluateAsDouble() to evaluatesToReal() and evaluateAsReal() respectively. Renamed Expression::set() to Expression::initialise(). Exposed Expression::initialise() and Expression::evaluatesToReal() in ExpressionVariable. Code compiles and links.

+++ src/variables/real.h
@@ -26,7 +26,7 @@
 #include "templates/vobject.h"
 
 // Real Variable
-class RealVariable : public Variable, VObject<int>
+class RealVariable : public Variable, VObject<double>
 {
 	public:
 	// Constructor
